Extract partition consumer state management code to a PartitionConsumingMapReduce. Allows user code to be less verbose.
Conflicts:
	cdap-unit-test/src/test/java/co/cask/cdap/partitioned/AppWithMapReduceConsumingPartitions.java
